ldap_scraper/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
ldap3==2.9